added  "about us " section
The about us section could be implemented by adding a new page to the website or by adding a section to the footer. The section could include the following information:
1.Company name
2.Mission statement

I believe that adding an about us section to the website would be a valuable addition that would benefit both the company and its customers.